backend/requirements.txt
cffi==1.15.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.14.0
flask-serialize==2.1.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
permissive-dict==1.0.4
pycparser==2.21
pydantic==1.10.5
PyJWT==2.7.0
pymssql==2.2.7
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pythonnet==3.0.1
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
